+++ routes/index.js
@@ -1,11 +1,9 @@
 const express = require('express')
-const { get } = require('../controllers/index')
 const transactionRoutes = require('./transactions.routes');
 
 const router = express.Router()
 
 // example of a route with index controller get function
-router.get('/', get)
-router.use('/transactions', transactionRoutes);
+router.use('/api/v1', transactionRoutes);
 
 module.exports = router
