+++ routes/transactions.routes.js
@@ -4,8 +4,12 @@ const postTransactionSchema = require('../schemas/transaction/postTransactionSch
 const {
     validateRequestSchema,
   } = require('../middlewares/validation/validate-schema.middleware');
+const { put } = require('../../controllers/transactions/')
+const { putValidation } = require('../schemas/transaction');
+
 const router = express.Router();
 
+router.put('/transactions/:id', validateRequestSchema(putValidation), put);
 router.post('/', validateRequestSchema(postTransactionSchema), postCreateTransaction);
 
-module.exports = router;
+module.exports = router
